+++ client/capi/base/c_api.cpp
@@ -3,7 +3,7 @@
 // Created Date: 02/07/2018
 // Author: Shun Suzuki
 // -----
-// Last Modified: 30/04/2021
+// Last Modified: 01/05/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2018-2020 Hapis Lab. All rights reserved.
@@ -42,14 +42,14 @@ bool AUTDOpenControllerWith(void* const handle, void* const p_link) {
   if (res.is_err()) LastError() = res.unwrap_err();
   return res.unwrap_or(false);
 }
-int32_t AUTDAddDevice(void* const handle, const autd::Float x, const autd::Float y, const autd::Float z, const autd::Float rz1, const autd::Float ry,
-                      const autd::Float rz2, const int32_t group_id) {
+int32_t AUTDAddDevice(void* const handle, const float x, const float y, const float z, const float rz1, const float ry, const float rz2,
+                      const int32_t group_id) {
   auto* cnt = static_cast<ControllerWrapper*>(handle);
   const auto res = cnt->ptr->geometry()->AddDevice(autd::Vector3(x, y, z), autd::Vector3(rz1, ry, rz2), group_id);
   return static_cast<int32_t>(res);
 }
-int32_t AUTDAddDeviceQuaternion(void* const handle, const autd::Float x, const autd::Float y, const autd::Float z, const autd::Float qua_w,
-                                const autd::Float qua_x, const autd::Float qua_y, const autd::Float qua_z, const int32_t group_id) {
+int32_t AUTDAddDeviceQuaternion(void* const handle, const float x, const float y, const float z, const float qua_w, const float qua_x,
+                                const float qua_y, const float qua_z, const int32_t group_id) {
   auto* cnt = static_cast<ControllerWrapper*>(handle);
   const auto res = cnt->ptr->geometry()->AddDeviceQuaternion(autd::Vector3(x, y, z), autd::Quaternion(qua_w, qua_x, qua_y, qua_z), group_id);
   return static_cast<int32_t>(res);
@@ -140,13 +140,13 @@ bool AUTDIsSilentMode(void* const handle) {
   auto* cnt = static_cast<ControllerWrapper*>(handle);
   return cnt->ptr->silent_mode();
 }
-autd::Float AUTDWavelength(void* const handle) {
+float AUTDWavelength(void* const handle) {
   auto* cnt = static_cast<ControllerWrapper*>(handle);
-  return cnt->ptr->geometry()->wavelength();
+  return static_cast<float>(cnt->ptr->geometry()->wavelength());
 }
-void AUTDSetWavelength(void* const handle, const autd::Float wavelength) {
+void AUTDSetWavelength(void* const handle, const float wavelength) {
   auto* cnt = static_cast<ControllerWrapper*>(handle);
-  return cnt->ptr->geometry()->set_wavelength(wavelength);
+  return cnt->ptr->geometry()->set_wavelength(static_cast<autd::Float>(wavelength));
 }
 int32_t AUTDNumDevices(void* const handle) {
   auto* cnt = static_cast<ControllerWrapper*>(handle);
@@ -246,13 +246,13 @@ void AUTDSequence(void** out) {
   auto* s = SequencePtrCreate(autd::sequence::PointSequence::Create());
   *out = s;
 }
-bool AUTDSequenceAddPoint(void* const seq, const autd::Float x, const autd::Float y, const autd::Float z) {
+bool AUTDSequenceAddPoint(void* const seq, const float x, const float y, const float z) {
   auto* seq_w = static_cast<SequenceWrapper*>(seq);
   auto res = seq_w->ptr->AddPoint(autd::Vector3(x, y, z));
   if (res.is_err()) LastError() = res.unwrap_err();
   return res.unwrap_or(false);
 }
-bool AUTDSequenceAddPoints(void* const seq, autd::Float* points, const uint64_t size) {
+bool AUTDSequenceAddPoints(void* const seq, float* points, const uint64_t size) {
   auto* seq_w = static_cast<SequenceWrapper*>(seq);
   std::vector<autd::Vector3> p;
   for (size_t i = 0; i < size; i++) p.emplace_back(autd::Vector3(points[3 * i], points[3 * i + 1], points[3 * i + 2]));
@@ -260,17 +260,17 @@ bool AUTDSequenceAddPoints(void* const seq, autd::Float* points, const uint64_t 
   if (res.is_err()) LastError() = res.unwrap_err();
   return res.unwrap_or(false);
 }
-autd::Float AUTDSequenceSetFreq(void* const seq, const autd::Float freq) {
+float AUTDSequenceSetFreq(void* const seq, const float freq) {
   auto* seq_w = static_cast<SequenceWrapper*>(seq);
-  return seq_w->ptr->SetFrequency(freq);
+  return static_cast<float>(seq_w->ptr->SetFrequency(freq));
 }
-autd::Float AUTDSequenceFreq(void* const seq) {
+float AUTDSequenceFreq(void* const seq) {
   auto* seq_w = static_cast<SequenceWrapper*>(seq);
-  return seq_w->ptr->frequency();
+  return static_cast<float>(seq_w->ptr->frequency());
 }
-autd::Float AUTDSequenceSamplingFreq(void* const seq) {
+float AUTDSequenceSamplingFreq(void* const seq) {
   auto* seq_w = static_cast<SequenceWrapper*>(seq);
-  return seq_w->ptr->sampling_frequency();
+  return static_cast<float>(seq_w->ptr->sampling_frequency());
 }
 uint16_t AUTDSequenceSamplingFreqDiv(void* const seq) {
   auto* seq_w = static_cast<SequenceWrapper*>(seq);
@@ -280,8 +280,8 @@ void AUTDDeleteSequence(void* const seq) {
   auto* seq_w = static_cast<SequenceWrapper*>(seq);
   SequenceDelete(seq_w);
 }
-void AUTDCircumSequence(void** out, const autd::Float x, const autd::Float y, const autd::Float z, const autd::Float nx, const autd::Float ny,
-                        const autd::Float nz, const autd::Float radius, const uint64_t n) {
+void AUTDCircumSequence(void** out, const float x, const float y, const float z, const float nx, const float ny, const float nz, const float radius,
+                        const uint64_t n) {
   auto* s = SequencePtrCreate(autd::sequence::CircumSeq::Create(autd::Vector3(x, y, z), autd::Vector3(nx, ny, nz), radius, n));
   *out = s;
 }
@@ -322,7 +322,7 @@ void AUTDAddSTMGain(void* const handle, void* const gain) {
   auto* g = static_cast<GainWrapper*>(gain);
   cnt->ptr->AddSTMGain(g->ptr);
 }
-bool AUTDStartSTModulation(void* const handle, const autd::Float freq) {
+bool AUTDStartSTModulation(void* const handle, const float freq) {
   auto* cnt = static_cast<ControllerWrapper*>(handle);
   auto res = cnt->ptr->StartSTModulation(freq);
   if (res.is_err()) LastError() = res.unwrap_err();
@@ -356,31 +356,25 @@ int32_t AUTDDeviceIdxForTransIdx(void* const handle, const int32_t global_trans_
   const auto res = cnt->ptr->geometry()->device_idx_for_trans_idx(global_trans_idx);
   return static_cast<int32_t>(res);
 }
-autd::Float* AUTDTransPositionByGlobal(void* const handle, const int32_t global_trans_idx) {
+void AUTDTransPositionByGlobal(void* const handle, const int32_t global_trans_idx, float* x, float* y, float* z) {
   auto* cnt = static_cast<ControllerWrapper*>(handle);
   const auto pos = cnt->ptr->geometry()->position(global_trans_idx);
-  auto* array = new autd::Float[3];
-  array[0] = pos.x();
-  array[1] = pos.y();
-  array[2] = pos.z();
-  return array;
+  *x = static_cast<float>(pos.x());
+  *y = static_cast<float>(pos.y());
+  *z = static_cast<float>(pos.z());
 }
-autd::Float* AUTDTransPositionByLocal(void* const handle, const int32_t device_idx, const int32_t local_trans_idx) {
+void AUTDTransPositionByLocal(void* const handle, const int32_t device_idx, const int32_t local_trans_idx, float* x, float* y, float* z) {
   auto* cnt = static_cast<ControllerWrapper*>(handle);
   const auto pos = cnt->ptr->geometry()->position(device_idx, local_trans_idx);
-  auto* array = new autd::Float[3];
-  array[0] = pos.x();
-  array[1] = pos.y();
-  array[2] = pos.z();
-  return array;
+  *x = static_cast<float>(pos.x());
+  *y = static_cast<float>(pos.y());
+  *z = static_cast<float>(pos.z());
 }
-autd::Float* AUTDDeviceDirection(void* const handle, const int32_t device_idx) {
+void AUTDDeviceDirection(void* const handle, const int32_t device_idx, float* x, float* y, float* z) {
   auto* cnt = static_cast<ControllerWrapper*>(handle);
   const auto dir = cnt->ptr->geometry()->direction(device_idx);
-  auto* array = new autd::Float[3];
-  array[0] = dir.x();
-  array[1] = dir.y();
-  array[2] = dir.z();
-  return array;
+  *x = static_cast<float>(dir.x());
+  *y = static_cast<float>(dir.y());
+  *z = static_cast<float>(dir.z());
 }
 #pragma endregion