+++ client/capi/gain_holo/gain_holo.h
@@ -3,15 +3,13 @@
 // Created Date: 20/02/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 05/04/2021
+// Last Modified: 01/05/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
 //
 
 #pragma once
-
-#include "autd_types.hpp"
 
 #if WIN32
 #define EXPORT_AUTD __declspec(dllexport)
@@ -20,5 +18,5 @@
 #endif
 
 extern "C" {
-EXPORT_AUTD void AUTDHoloGain(void** gain, autd::Float* points, autd::Float* amps, int32_t size, int32_t method, void* params);
+EXPORT_AUTD void AUTDHoloGain(void** gain, float* points, float* amps, int32_t size, int32_t method, void* params);
 }