+++ client/examples/examples/stm.hpp
@@ -21,11 +21,11 @@ inline void STMTest(const autd::ControllerPtr& autd) {
   const auto m = autd::modulation::Modulation::Create(255);
   autd->AppendModulationSync(m).unwrap();
 
-  const auto center = autd::Vector3(TRANS_SIZE_MM * ((NUM_TRANS_X - 1) / 2.0f), TRANS_SIZE_MM * ((NUM_TRANS_Y - 1) / 2.0f), 150.0f);
+  const auto center = autd::Vector3(TRANS_SIZE_MM * ((NUM_TRANS_X - 1) / 2.0), TRANS_SIZE_MM * ((NUM_TRANS_Y - 1) / 2.0), 150.0);
   const auto point_num = 200;
   for (auto i = 0; i < point_num; i++) {
-    const auto radius = 30.0f;
-    const auto theta = 2.0f * PI * static_cast<float>(i) / point_num;
+    const auto radius = 30.0;
+    const auto theta = 2.0 * PI * static_cast<autd::Float>(i) / point_num;
     const auto pos = autd::Vector3(radius * cos(theta), radius * sin(theta), 0.0);
     const auto g = autd::gain::FocalPointGain::Create(center + pos);
     autd->AddSTMGain(g);