+++ client/capi/link_twincat/c_api.cpp
@@ -3,13 +3,11 @@
 // Created Date: 02/07/2018
 // Author: Shun Suzuki
 // -----
-// Last Modified: 04/04/2021
+// Last Modified: 02/05/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2018-2020 Hapis Lab. All rights reserved.
 //
-
-#include <cstdint>
 
 #include "../base/wrapper_link.hpp"
 #include "./twincat_link.h"