+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Make double-precision float default
     * `USE_DOUBLE_AUTD` was removed, and `USE_SINGLE_FLOAT_AUTD` was added
     * capi still use single-precision float in its interface, but it uses double internally
+* Fix issue #2
 
 # 0.9.1
 * Fix bugs in `autdsoem`